Add support for the attribute servers of Cve_announcements
